1.1
Изменено
1. Переименован файл изменяющий профайл для лучшего понимания его функционала
2. Удаление пользователей из базы данных исправлено, теперь пользователи удаляются корректно
3. Добавлена возможность удаления пользователей через клавишу делит в списке пользователей
4. Ограничена возможность удаление текущего пользователя администратора
5. Исправлено сообщение, возникающее при сохранении пользователя с пустыми ФИО
6. Увеличен интервал запроса обновления списка пользователей с 2 секунд до 5
7. Для быстродействия некоторые методы помещены noexcept
8. В визарде изменено:
8.1. Добавлен гуй для сервиса папок и файлов теперь сохраняет пути и поддерживает добавление и удаление из списка
8.2 Гуй сервиса папок позволяет добавлять у удалять сторонние папки и файлы, в папках отображается количество файлов в папках.
8.2. Гуй для переменных окружения поддерживает изменения полей, удаление и отмену удаления через Cntr+Z
8.3. В таблицах гуя переменных окружения выставлены правильные размеры колонок, теперь скролл бар отсутствует
8.4. Надпись переменные окружения в гуе визаржа отцентрована
8.5. Обводка листВиджета изменена для большей наглядности на белую
8.6. Добавлен тултип для новых гуев визарда
9. В сигнале удаления пользователей убраны ссылки
10. В поле ЮзерЕдит мессадж бокс статический
11. Убраны неиспользуемые стили сплиттера

+++ Admin_GUI/Views/admingui.cpp
@@ -74,6 +74,7 @@ void Admin_GUI::InsertWidgetsIntoLayout()
 void Admin_GUI::ConnectObjects()
 {
     connect(m_usersListWidget, &LinuxUsersListWidget::ToUserClick, this, &Admin_GUI::OnUserClick);
+    connect(m_usersListWidget, &LinuxUsersListWidget::ToDeleteUser, this, &Admin_GUI::OnDeleteUser);
     connect(m_userEditPanel, &UserEditPanel::ToRoleToViewChanged, m_roleEditPanel, &RoleEditPanel::OnRoleToViewChanged);
     connect(m_userEditPanel, &UserEditPanel::ToDeleteUser, this, &Admin_GUI::OnDeleteUser);
     connect(m_userEditPanel, &UserEditPanel::ToSaveUser, this, &Admin_GUI::OnSaveUser);
@@ -109,12 +110,19 @@ void Admin_GUI::OnUserClick(const User &user)
 void Admin_GUI::OnDeleteUser(const QString &userId, const QString &userName)
 {
     const int roleId = m_userModel->GetRoleIdByUserId(userId);
-    m_userModel->DeleteUser(userId, userName);
-    if (User::RoleIsValid(roleId))
+    if(userId==m_linuxUserService->GetCurrentUserId() && userName==m_linuxUserService->GetCurrentUserName())
     {
-        m_userDesktopService->DeleteAllIconsToUser(roleId, userName);
+        QMessageBox::warning(this, QStringLiteral("Удаление пользователей"), QStringLiteral("Нельзя удалить текущего пользователя - администратора"));
     }
-    m_userDesktopService->GetAllUserDesktops(userName);
+    else
+    {
+        m_userModel->DeleteUser(userId, userName);
+        if (User::RoleIsValid(roleId))
+        {
+            m_userDesktopService->DeleteAllIconsToUser(roleId, userName);
+        }
+        m_userDesktopService->GetAllUserDesktops(userName);
+    }
 }
 
 void Admin_GUI::OnSaveUser(const QString &userId, const QString &userName, const QString &FCS, int oldRole, int newRole)
